chore: split large source files and clean up TODO
Extract test modules and slim down oversized files under the ~800-line target.

- docs/todo.md: 149 → 74 lines, remove completed items, consolidate
- crates/sandbox: extract tests to tests.rs (lib.rs: 1182 → 612)
- tui/render/bottom_pane: extract tests to bottom_pane_tests.rs (914 → 556)

Net: -1006 lines in tracked source files.
